requirements.txt.txt
streamlit
torch
torchvision
numpy
pillow
scikit-learn
matplotlib
